seperate colours class
