+++ Momento/extensions/cogs/moderation.py
@@ -12,19 +12,7 @@ from discord.ext.commands.converter import Converter, MemberConverter
 from discord.ext.commands.errors import BadArgument, CheckFailure
 from discord.member import Member
 from discord.message import Message
-
-
-class pcolours:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKCYAN = '\033[96m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-
+from classes import pcolours
 
 
 profanity.load_censor_words()
