+++ Momento/extensions/cogs/player-stats-tracker.py
@@ -8,18 +8,7 @@ from discord.ext.commands import Cog
 from discord import utils
 from datetime import datetime
 import math
-
-
-class pcolours:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKCYAN = '\033[96m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
+from classes import pcolours
 
 
 class PlayerStatsTracker(Cog):
