+++ Momento/extensions/cogs/welcome.py
@@ -7,19 +7,9 @@ from discord.ext.commands.cog import Cog
 from discord.utils import get
 from db import db
 import datetime
+from classes import pcolours
 
 
-
-class pcolours:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKCYAN = '\033[96m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
 
 
 class Welcomer(Cog):
